+++ packages/el-protable/vite.config.ts
@@ -5,24 +5,17 @@ import { resolve } from "path";
 import dts from "vite-plugin-dts";
 export default defineConfig({
 	plugins: [
-		// @ts-ignore
 		dts({
 			copyDtsFiles: true,
 			outDir: [
 				"dist",
-				// "types",
-				// 'types/inner'
 			],
 			clearPureImport: false,
 			include: ["src/*", "src/types.ts"],
-			// staticImport: true,
-			// rollupTypes: true,
 			insertTypesEntry: true,
 			compilerOptions: {
-				// declarationMap: true,
 			},
 		}),
-		// @ts-ignore
 		vueJsx(),
 		vue(),
 	],