+++ packages/naive-protable/src/index.vue
@@ -1,9 +1,16 @@
-<template> </template>
+<template>
+  <div class="n-protable">
+    <div class="n-protable-search">
+      <Search></Search>
+    </div>
+  </div>
+</template>
 
 <script setup lang="ts">
+import Search from "./components/Search.vue";
 defineOptions({
-  name:"NProTable"
-})
+  name: "NProTable",
+});
 </script>
 
 <style scoped lang="scss"></style>