+++ CHANGELOG.md
@@ -2,6 +2,22 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [0.0.4](https://github.com/denganjia/suite-kit/compare/v0.0.3...v0.0.4) (2024-11-18)
+
+
+### Features
+
+* 新增功能 ([6f2fcee](https://github.com/denganjia/suite-kit/commit/6f2fceea6d75bbc4abbd870f143e4a6cd5834968))
+* 增加NProTable中 ([489dc84](https://github.com/denganjia/suite-kit/commit/489dc8460ec957ae724fc021ea4046cfb24ec4ab))
+* release @suite-kit/el-protable@0.0.4 ([5b07f89](https://github.com/denganjia/suite-kit/commit/5b07f89c415141bea4a2ea520faf698605414c44))
+* release @suite-kit/el-protable@0.0.5 ([3b8ec8c](https://github.com/denganjia/suite-kit/commit/3b8ec8ccbd799e642ce96fb7fd8411dc7578cabd))
+
+
+### Bug Fixes
+
+* 修复操作列“operation”可以控制显示隐藏的BUG ([3a4ef40](https://github.com/denganjia/suite-kit/commit/3a4ef4063d4037fa2dec68232df45aa3a96f9adf))
+* 修改BUG ([42d054d](https://github.com/denganjia/suite-kit/commit/42d054dc6d1c8b9ad3c591ecf04f4f5dc226e203))
+
 ### [0.0.3](https://github.com/denganjia/suite-kit/compare/v0.0.2...v0.0.3) (2023-09-08)